Updates v1.123
- Fixed nil pointer bug when delete file during storage quota is updating
- Fixed force delete lag bug when deleting a lot of large files
- Fixed File Manager upload counter and waiting counter bug
- Fixed Music App cannot playback filename with hashtag bug
- Fixed FTP bug in loading in virtual file system handler
- Fixed large file scanner bug
- Fixed File Manager cannot open file with two dot bug
- Fixed Basic backup cache file bug
- Fixed shortcut app selector bug
- Fixed AGI file glob and aglob showing hidden file bug
- Fixed file operation complete checkmark not shown bug
- Fixed desktop clone file bug
- Fixed NotepadA load json bug
- Fixed file operation dialog bug
- Fixed HDSv2 execute URL escape bug
- Fixed WebApp reload on non-admin account bug
- Fixed Linux netstat bug
- Added Screen Capture function
- Added Ctrl + C shortcut for desktop file copy
- Added experimental upload mode for uploading huge file
- Added thumbnail in Music WebApp song list
- Added shortcut on Mobile Desktop Interface
- Added GET paramter and hash passthrough in login redirection interface
- Added File Manager "unzip here" feature
- Added localization to shortcut creation interface
- Added desktop admin icon after username
- Added desktop Ctrl+C and Ctrl+V support
- Added file versioning system
- Added Makefile
- Added Expotential login retry time
- Added no-title placeholder for MangaCafe
- Added enable overview in non-cloud mode
- Added auto tar.gz unzip logic during startup
- Added netstat support for MacOS (By jk)
- Optimized group unzip logic
- Optimized code in file operation dialog
- Optimized LDAP logic (By Alanyeung)
- Migrated all high abstraction code base to use common module
- Moved all trash, cache and versioning file to .metadata folder

+++ src/mod/filesystem/fileOpr.go
@@ -117,34 +117,40 @@ func ArozUnzipFileWithProgress(filelist []string, outputfile string, progressHan
 			defer reader.Close()
 			path := filepath.Join(outputfile, file.Name)
 
-			err = os.MkdirAll(path, os.ModePerm)
+			parentFolder := path
+			if !file.FileInfo().IsDir() {
+				//Is file. Change target to its parent dir
+				parentFolder = filepath.Dir(path)
+			}
+
+			err = os.MkdirAll(parentFolder, 0775)
 			if err != nil {
 				return err
 			}
 
 			if file.FileInfo().IsDir() {
-				//Folder extracted
-
+				//Folder is created already be the steps above.
 				//Update the progress
 				unzippedFileCount++
 				progressHandler(file.Name, unzippedFileCount, totalFileCounts, float64(unzippedFileCount)/float64(totalFileCounts)*100.0)
 				continue
 			}
 
-			err = os.Remove(path)
-			if err != nil {
-				return err
-			}
-
+			//Extrat and write to the target file
 			writer, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, file.Mode())
 			if err != nil {
 				return err
 			}
-			defer writer.Close()
 			_, err = io.Copy(writer, reader)
 			if err != nil {
+				//Extraction failed. Remove this file if exists
+				writer.Close()
+				if FileExists(path) {
+					os.Remove(path)
+				}
 				return err
 			}
+			writer.Close()
 
 			//Update the progress
 			unzippedFileCount++
